refactor: restructure api/ to match professor's spec
