--- myapp/api/Database/__init__.py
@@ -1,1 +1,0 @@
-from .database import get_db, SessionLocal, engine, DATABASE_URL
