--- myapp/api/routers/__init__.py
@@ -1,1 +1,0 @@
-# Pulse API — Router package
