--- myapp/api/schemas/__init__.py
@@ -1,1 +1,0 @@
-# Pulse API — Pydantic schemas package
